Move mesh comparison test to common pytest code
