+++ setup.cfg
@@ -8,6 +8,7 @@ source = nanomesh
 [tool:pytest]
 testpaths = tests
 addopts = --cov --cov-report xml --cov-report term --cov-report html
+norecursedirs=tests/helpers
 
 # Define `python setup.py build_sphinx`
 [build_sphinx]
